feat: Comment and Review

+++ source/src/main/java/com/tenant/api/controller/ReviewController.java
@@ -4,14 +4,18 @@ import com.tenant.api.constant.BaseConstant;
 import com.tenant.api.dto.ApiMessageDto;
 import com.tenant.api.dto.ErrorCode;
 import com.tenant.api.dto.ResponseListDto;
+import com.tenant.api.dto.reaction.VoteDto;
 import com.tenant.api.dto.review.ReviewDto;
+import com.tenant.api.dto.review.ReviewStatisticsDto;
 import com.tenant.api.exception.BadRequestException;
 import com.tenant.api.exception.NotFoundException;
 import com.tenant.api.exception.UnauthorizationException;
+import com.tenant.api.form.ChangeStatusForm;
 import com.tenant.api.form.reaction.CreateReactionForm;
 import com.tenant.api.form.review.CreateReviewForm;
 import com.tenant.api.form.review.UpdateReviewForm;
 import com.tenant.api.mapper.ReviewMapper;
+import com.tenant.api.service.MovieService;
 import com.tenant.api.storage.tenant.criteria.ReviewCriteria;
 import com.tenant.api.storage.tenant.model.Movie;
 import com.tenant.api.storage.tenant.model.Reaction;
@@ -56,22 +60,32 @@ public class ReviewController extends ABasicController {
     @Autowired
     private ReactionRepository reactionRepository;
 
+    @Autowired
+    private MovieService movieService;
+
     @Transactional("tenantTransactionManager")
     @PostMapping(value = "/create", produces = MediaType.APPLICATION_JSON_VALUE)
-    @PreAuthorize("hasRole('REV_C')")
-    public ApiMessageDto<Void> create(@Valid @RequestBody CreateReviewForm form) {
+//    @PreAuthorize("hasRole('REV_C')")
+    public ApiMessageDto<ReviewDto> create(@Valid @RequestBody CreateReviewForm form) {
         Movie movie = movieRepository.findById(form.getMovieId())
                 .orElseThrow(() -> new BadRequestException("[Movie] not found", ErrorCode.MOVIE_ERROR_NOT_FOUND));
 
         User user = userRepository.findById(getCurrentUser())
                 .orElseThrow(() -> new NotFoundException("[User] not found", ErrorCode.USER_ERROR_NOT_FOUND));
 
+        if (reviewRepository.existsByAuthorIdAndMovieId(user.getId(), movie.getId())) {
+            throw new BadRequestException("[Review] already existed", ErrorCode.REVIEW_ERROR_EXISTED);
+        }
+
         Review review = reviewMapper.fromCreateReviewFormToEntity(form);
         review.setAuthor(user);
         review.setMovieId(movie.getId());
         reviewRepository.save(review);
 
-        return makeSuccessResponse("Create review success");
+        ReviewStatisticsDto statistics = movieService.calculateReview(movie.getId(), review.getRate(), BaseConstant.ACTION_ADD);
+        ReviewDto reviewDto = reviewMapper.entityToReviewDto(review);
+        reviewDto.setStatistics(statistics);
+        return makeSuccessResponse(reviewDto, "Create review success");
     }
 
     @GetMapping(value = "/get/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -88,7 +102,7 @@ public class ReviewController extends ABasicController {
                 pageable.getPageSize(),
                 Sort.by(Sort.Order.desc("rate"), Sort.Order.desc("createdDate")));
 
-        criteria.setStatus(BaseConstant.STATUS_ACTIVE);
+//        criteria.setStatus(BaseConstant.STATUS_ACTIVE);
         Page<Review> reviews = reviewRepository.findAll(criteria.getSpecification(), pageable);
         return makeSuccessResponse(makeResponseListDto(reviews, reviewMapper::fromEntityToReviewDtoList), "Get list review success");
     }
@@ -166,13 +180,41 @@ public class ReviewController extends ABasicController {
 
     @Transactional("tenantTransactionManager")
     @DeleteMapping(value = "/delete/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    @PreAuthorize("hasRole('REV_D')")
-    public ApiMessageDto<Void> delete(@PathVariable("id") Long id) {
+//    @PreAuthorize("hasRole('REV_D')")
+    public ApiMessageDto<ReviewStatisticsDto> delete(@PathVariable("id") Long id) {
         Review review = reviewRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("[Review] Not found", ErrorCode.REVIEW_ERROR_NOT_FOUND));
 
+        if (!isUser() && !Objects.equals(review.getAuthor().getId(), getCurrentUser())) {
+            throw new UnauthorizationException("Not allow");
+        }
+
         reactionRepository.deleteByReviewId(review.getId());
         reviewRepository.delete(review);
-        return makeSuccessResponse("Delete review success");
+
+        ReviewStatisticsDto statistics = movieService.calculateReview(review.getMovieId(), review.getRate(), BaseConstant.ACTION_DELETE);
+        return makeSuccessResponse(statistics, "Delete review success");
+    }
+
+    @GetMapping(value = "/vote-list/{movieId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ApiMessageDto<List<VoteDto>> voteList(@PathVariable("movieId") Long movieId) {
+        return makeSuccessResponse(reviewRepository.findVotesByMovieIdAndUserId(movieId, getCurrentUser()), "Get list vote success");
+    }
+
+    @Transactional("tenantTransactionManager")
+    @PutMapping(value = "/change-status", produces = MediaType.APPLICATION_JSON_VALUE)
+    @PreAuthorize("hasRole('REV_C_S')")
+    public ApiMessageDto<Void> changeStatus(@Valid @RequestBody ChangeStatusForm form) {
+        Review review = reviewRepository.findById(form.getId())
+                .orElseThrow(() -> new NotFoundException("[Review] Not found", ErrorCode.COMMENT_ERROR_NOT_FOUND));
+        review.setStatus(form.getStatus());
+        reviewRepository.save(review);
+        return makeSuccessResponse("Change status success");
+    }
+
+    @GetMapping(value = "/check/{movieId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ApiMessageDto<ReviewDto> check(@PathVariable Long movieId) {
+        Review review = reviewRepository.findByAuthorIdAndMovieId(getCurrentUser(), movieId).orElse(null);
+        return makeSuccessResponse(reviewMapper.entityToReviewDto(review), "Get review success");
     }
 }
